Переход на новый фреймворк. Добавил тесты на эндпоинты депозит и трансфер 2

+++ src/main/java/models/TransferMoneyRequest.java
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @Builder
 
 public class TransferMoneyRequest extends BaseModel {
-    private int senderAccountId;
-    private int receiverAccountId;
+    private long senderAccountId;
+    private long receiverAccountId;
     private Number amount;
 }
